+++ src/app/cart/cart.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
 import { CartComponentComponent } from './components/cart-component/cart-component.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
@@ -11,7 +10,6 @@ import { SharedModule } from './../shared/shared.module';
     CartItemComponent
   ],
   imports: [
-    CommonModule,
     SharedModule
   ],
   exports: [