+++ src/app/cart/components/cart-component/cart-component.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CartService } from '../../services/cart.service';
 import { CartItem } from '../../models/cart-item.model';
+import { Field } from '../../models/field.model';
 
 @Component({
   selector: 'app-cart-component',
@@ -9,12 +10,23 @@ import { CartItem } from '../../models/cart-item.model';
 })
 export class CartComponentComponent implements OnInit {
 
+  selectedField: Field = new Field('name', 'Name');
+  selectedDirection: string;
+  fields = [
+     new Field('name', 'Name' ),
+     new Field('count', 'Count' ),
+     new Field('price', 'Price' ),
+     new Field('updated', 'Updated' )
+  ];
+
   cart: Array<CartItem>;
   yourAddress = '';
 
   constructor(private cartService: CartService) { }
 
   ngOnInit() {
+    this.selectedField.id = 'name';
+    this.selectedDirection = 'true';
     this.getCart();
   }
 
@@ -51,5 +63,14 @@ export class CartComponentComponent implements OnInit {
     console.log('addAddress event');
     this.yourAddress = value;
   }
+
+  onSelect(fieldId: string) {
+    this.selectedField = null;
+    this.fields.forEach(field => {
+      if (field.id === fieldId) {
+        this.selectedField = field;
+      }
+    });
+}
 
 }