+++ src/app/layout/layout.module.ts
@@ -1,13 +1,11 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { AboutComponent } from './components/about/about.component';
-
-
+import { SharedModule } from './../shared/shared.module';
 
 @NgModule({
   declarations: [AboutComponent],
   imports: [
-    CommonModule
+    SharedModule
   ],
   exports: [
     AboutComponent