+++ src/app/products/models/product.model.ts
@@ -2,14 +2,6 @@ import { Category } from '../../enums';
 
 export class Product {
     constructor(
-    //   id, name, description, price, category, isAvailable) {
-    //     this.id = id;
-    //     this.name = name;
-    //     this.description = description;
-    //     this.price = price;
-    //     this.category = category;
-    //     this.isAvailable = isAvailable;
-    // }
         public id: number,
         public name: string,
         public description: string,