+++ src/app/shared/shared.module.ts
@@ -1,15 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 import { StyleDirective } from './directives/style.directive';
 import { ClickDirective } from './directives/click.directive';
-
-
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 @NgModule({
-  declarations: [StyleDirective, ClickDirective],
+  declarations: [StyleDirective, ClickDirective, OrderByPipe],
   imports: [
-    CommonModule
+    CommonModule, FormsModule
   ],
-  exports: [StyleDirective, ClickDirective]
+  exports: [StyleDirective, ClickDirective, OrderByPipe, CommonModule, FormsModule]
 })
 export class SharedModule { }